Refactor : 입력값 파싱과 예외처리를 서로 다른 두 객체로 분류
- ParsingUtil.java의 예외처리 메서드를 InputException 클래스로 생성하였다.
